Move make_vcs_requirement_url() to vcs/__init__.py.
